The pitcher can now reliably hold the ball and release the ball so that the ball retains its momentum
Created a ragdoll with the character, made all the character's bones' rigidbodies kinematic save for the ones in the arm, gave the ball a fixed joint between it and the hand. The fixed joint is destroyed when the user presses the space bar, thus releasing the ball and allowing the ball to retain its momentum, thus enabling an actual throw to occur.