Added a LearningManager script to handle the assignment of the parameters
At this point, the parameters consist of an array of Vector3s, with each Vector3 being the torque that is applied to a specific joint of the arm throughout the particular simulation, and an int that is the number of frames that should pass before the ball is thrown. Currently, when you click play, the simulation assigns random values for these parameters each time it resets itself (whenever the ball hits the ground).